requirements.txt

flask
flask-cors
python-dotenv
google-genai
pypdf
numpy
pandas
scikit-learn
matplotlib
python-pptx
gunicorn
